requirements.txt
aniso8601==8.1.1
autopep8==1.5.5
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
flake8==3.8.4
Flask==1.1.2
Flask-Cors==3.0.10
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
six==1.15.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==1.0.1
